remove useEnsured hook
This is theoretically useful, but we have phased it out in Garden because it adds an extra network request to "ensure" a resource exists on every fresh load of a page that uses it.

I think it's generally better to handle resources not existing when an error actually occurs and to help the user resolve this situation rather than just blindly ensuring a resource exists, so remove this for now - open to adding it back in the future.